+++ PotreeConverter/include/PointReader.h
@@ -23,7 +23,7 @@ public:
 
 	virtual AABB getAABB() = 0;
 
-	virtual long numPoints() = 0;
+	virtual long long numPoints() = 0;
 
 	virtual void close() = 0;
 };
